Update WeCom version
WeCom version numbers were updated in the code. This is to make sure that the applications in use are up-to-date.

In the README.md file, a new section was added to provide a solution for reported issue regarding missing 'libiconv.dll' module.

During the task of sending group messages in WeCom, an improvement was made to reduce redundant task of closing existing message manager windows every time a group message is sent. Now, it only pops up the user information menu to access the message manager button which is considerably more direct and efficient.